Fix scaling scenario (#28)
Create VnodeRepairStates for current vnodes and then only update
repaired at for them. This way we don't keep track of old vnodes
based on the old topology.

This also means that repair state might be reset for the new vnodes
as we will then use the old repaired at from the whole table rather
than the individual vnodes.