Split ranges during repair
Added a repair configuration for "target" repair size. This works
with an assumption of uniform data distribution between partition
keys within the table.

When repairing a table of 1GB with a target size of 100MB we will
perform *at least* 10 repairs. This is done by calculating the
size of the table and the total number of tokens. With this
information we can calculate a "target tokens" per repair by:
target tokens = total tokens / (table size / target size)

This means that tables with less data will perform larger repair
sessions and tables with more data will perform smaller repair
sessions.