Summarize token ranges
Instead of looking for full vnodes in VnodeRepairStateFactoryImpl
we now look for any sub range that is covered by a vnode.

When building up the repair history in VnodeRepairStates we run it
through the summarizer which first transforms the sub ranges into
a NormalizedRange.

The normalized ranges are transformed in such a way that the start
token is always smaller than the end token.
Basically a vnode (5, 10] and a sub range (6, 7] becomes:
(MIN, MIN+5] and (MIN+1, MIN+2]. This makes it much easier to work
with ranges that are "wrapping" around Long.MAX_VALUE. Instead we can
simply work based on the assumption that start is always smaller than
end.

The summarizer then creates a list of the "full" normalized vnode
with a default repaired at, and then all the sub ranges and when
they were repaired.
Then it goes through and splits any overlapping ranges using an
appropriate timestamp based on the two ranges that are overlapping.
Then it merges adjacent sub ranges based on a merge strategy, where
the default checks if the ranges are repaired within one hour of each
other.