Build ecChronos With Java 17 - Issue #607 (#617)
* Build ecChronos With Java 17 - Issue #607

In order to provide support for Java 17 in EcChronos,
identify the minimal changes necessary for the project to
compile and run with Java 17. Additionally, implement these
changes and any necessary updates to Maven packages.

This change must be merged just after a release of ecChronos
with the Issue #616 done.

Closes #607

* Build ecChronos With Java 17 - Issue #607

In order to provide support for Java 17 in EcChronos,
identify the minimal changes necessary for the project to
compile and run with Java 17. Additionally, implement these
changes and any necessary updates to Maven packages.

This change must be merged just after a release of ecChronos
with the Issue #616 done.

Closes #607

* Build ecChronos With Java 17 - Issue #607

In order to provide support for Java 17 in EcChronos,
identify the minimal changes necessary for the project to
compile and run with Java 17. Additionally, implement these
changes and any necessary updates to Maven packages.

This change must be merged just after a release of ecChronos
with the Issue #616 done.

Closes #607

* Build ecChronos With Java 17 - Issue #607

In order to provide support for Java 17 in EcChronos,
identify the minimal changes necessary for the project to
compile and run with Java 17. Additionally, implement these
changes and any necessary updates to Maven packages.

This change must be merged just after a release of ecChronos
with the Issue #616 done.

Closes #607

* Update surefile plugin

* Rebase Branch with Master

* Add Actions to Check different Java versions

* Fix Java 21 failures